Rename prismic-configuration.js to prismic.js
