+++ pages/[uid].js
@@ -1,6 +1,6 @@
 import { SliceZone } from "@prismicio/react";
 import * as prismicH from "@prismicio/helpers";
-import { createClient } from "../prismic-configuration";
+import { createClient } from "../prismic";
 import resolver from "../sm-resolver.js";
 import { linkResolver } from "../utils/linkResolver";
