+++ pages/_app.js
@@ -2,9 +2,8 @@ import "../styles/index.css";
 
 import React from "react";
 import NextApp from "next/app";
-import Script from "next/script";
+import { apiEndpoint } from "./../prismic"; // import the endpoint name from where it's defined
 
-import { apiEndpoint } from "./../prismic-configuration"; // import the endpoint name from where it's defined
 const prismicRepoName = /([a-zA-Z0-9-]+)?(\.cdn)?\.prismic\.io/.exec(
   apiEndpoint
 )[1]; //Regex to get repo ID
